neo4j/requirements.txt
boto3==1.16.4
neo4j-driver==4.1.1
jsonschema==3.2.0             
simplejson==3.16.0
PyJWT==1.7.1  
